requirements.txt
google-generativeai==0.8.4
flask==3.1.0
python-dotenv==1.1.0
chromadb==1.0.3
